refactor(visual): split scene scaffold renderers
- Move scene and data-band Code Studio renderer bodies into a focused renderer module behind the existing registry.

- Keep the template scaffold module responsible for spec selection, shared shell helpers, and wrapper registration.

- Update cleanup audit and codebase map with the new renderer boundary.
